[backport] Move dummyNCC to ncclient package
This change places dummyNCC implementation into `ncclient` package to
make the package boundary clear.

Also exposes dummyLBInterface struct details for testing purposes.
